# public/api/doc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BileMo REST API Documentation",
        "description": "REST API Documentation of BileMo",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://localhost:8000/",
            "description": "BileMo REST API"
        }
    ],
    "paths": {
        "/api/login_check": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Authentication",
                "description": "Authentication token access",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "T.Wellick"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "19N1t10n%"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK - Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJpYXQiOjE2MzI4NDI2NTgsImV4cCI6MTYzMjg0NjI1OCwicm9sZXMiOlsiUk9MRV9DVVNUT01FUiJdLCJ1c2VybmFtZSI6IlBzYUZyYW5jZSJ9.lpgfnOpZQ13Q-HtOCqtRdD8ChcI3NlWpTsiK9Ff6LlXiC7ih89tvxyPgj5xKeYFQQZlLS7838ukT_QnWdjbyEmG8NvZD5mD-jmMKBTGaEGMiVs8c0J7OcrFb_duKe8-9cmho3j1DvODzVXOqDTYL3J-C-2Qg0bW4RzCT1rVxgvAcFGHG5MbpFyvwft96V-ZablsmTE_7a3rkFnc8mvcWW4ivVr2UDKxcZKa9dQWXO72KMKF7-7eXEFmJSSXfRc9kUo-rXA9IWnJ9upzXWQeDP-0Ccgmzl6ukMzowUS2NsYwaHq7cmhB4wCTqTA8ubgpjYWbZiTxts0rbOjY5KW_0plB7VPrIarBSmi4rmjK0RJ68-MDrofpyMGy32S_TnBM0xN4cNqozxL0Rx9OxvyYnaxxSZ8NFOFQfABAevulCCG68jhXcEA6qknJ6OXVD3zESbFNYbkvt8iO48CuUXoRVb7Xv-mokrJW9CQpA7Aiz2M7vXYbEAIfux012Y_ZnTfDPseikYYT8xKFEGOFZ-oeejx15y7GbF5VLhV-IEisYC7Z7rTdsaa6OXt7sfJ1Ux9Y98bksWqncgOBddUwbGC9-r2cWF2V0C8-zcjHsfNfP5jttHgpbsEMguguX_al6QARYRnUu3_CrGqgQD2mh8EO4AX3EaswfuV-6IldyL7z3A7A"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/LoginUnauthorized"
                    },
                    "400": {
                        "$ref": "#/components/responses/LoginBadRequest"
                    }
                },
                "security": []
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Products index",
                "description": "List of products",
                "operationId": "7039000cbbf7804803dd7d52ca445c39",
                "parameters": [
                    {
                        "name": "brand",
                        "in": "query",
                        "description": "Refine to a specific brand",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Define an order sort",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number to ask",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductIndexList"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/ProductBadRequest"
                    },
                    "403": {
                        "$ref": "#/components/responses/ProductActionForbidden"
                    },
                    "401": {
                        "$ref": "#/components/responses/AccessUnauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/products/{productId}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Product detail",
                "description": "Detail of a product",
                "operationId": "e5929b87ce94317f33e78a7b84a91419",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "Id of asked product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK - Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductRead"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ProductNotFound"
                    },
                    "403": {
                        "$ref": "#/components/responses/ProductActionForbidden"
                    },
                    "401": {
                        "$ref": "#/components/responses/AccessUnauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductIndexList": {
                "description": "List of normalized product index",
                "properties": {
                    "meta": {
                        "properties": {
                            "totalPaginatedItems": {
                                "type": "integer",
                                "example": "3"
                            },
                            "maxItemsPerPage": {
                                "type": "integer",
                                "example": "5"
                            },
                            "lastPage": {
                                "type": "integer",
                                "example": "1"
                            },
                            "currentPage": {
                                "type": "integer",
                                "example": "1"
                            },
                            "currentPageItems": {
                                "type": "integer",
                                "example": "3"
                            },
                            "availableBrands": {
                                "type": "array",
                                "example": [
                                    "Apple",
                                    "Samsung",
                                    "Huawei",
                                    "Xiaomi"
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductIndex"
                        },
                        "example": [
                            {
                                "id": 1,
                                "model": "iPhone 12 - 256",
                                "brand": {
                                    "id": 1,
                                    "name": "Apple"
                                },
                                "_links": {
                                    "self": {
                                        "href": "https://localhost:8000/api/products/1"
                                    }
                                }
                            },
                            {
                                "id": 14,
                                "model": "P40 5G",
                                "brand": {
                                    "id": 3,
                                    "name": "Huawei"
                                },
                                "_links": {
                                    "self": {
                                        "href": "https://localhost:8000/api/products/14"
                                    }
                                }
                            },
                            {
                                "id": 10,
                                "model": "S21 5G - 256 Go - 8 Go",
                                "brand": {
                                    "id": 2,
                                    "name": "Samsung"
                                },
                                "_links": {
                                    "self": {
                                        "href": "https://localhost:8000/api/products/10"
                                    }
                                }
                            }
                        ]
                    }
                },
                "type": "object"
            },
...
            "contentResponseError": {
                "properties": {
                    "code": {
                        "type": "int"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "LoginUnauthorized": {
                "description": "Unauthorized",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/contentResponseError"
                        },
                        "example": {
                            "code": 401,
                            "message": "Informations de connexion non valides"
                        }
                    }
                }
            },
            "LoginBadRequest": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/contentResponseError"
                        },
                        "example": {
                            "code": 400,
                            "message": "La requête ne peut être traitée correctement"
                        }
                    }
                }
            },
...
        },
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}